Send CSRF token saved in cookie with every API call
