+++ frontend/src/subSites/participant/ParticipantExercise.js
@@ -27,6 +27,7 @@ export default class ParticipantExercise extends React.Component {
         this.handleSubmit = FormHandler.handleExerciseSubmit.bind(this);
         this.handleChange = FormHandler.handleChange.bind(this);
         this.postData = APIHandler.postData.bind(this);
+        this.getJSONHeader = APIHandler.getJSONHeader;
     };
 
     render() {
